feat: convenit to add instructions
