+++ .gitignore
@@ -54,4 +54,4 @@ dkms.conf
 #build directory
 build/
 dumps/
-tests/
+#tests/
