+++ README.md
@@ -1,0 +1,23 @@
+# x86-mips 二进制翻译器
+
+## Git Commit 格式
+
+```bash
+git commit -m "type: descriptions"
+```
+type 有以下种类：
+
+  * feat：新功能（feature）
+  * fix：修补bug
+  * docs：文档（documentation）
+  * style： 格式（不影响代码运行的变动）
+  * refactor：重构（即不是新增功能，也不是修改bug的代码变动）
+  * test：增加测试
+  * chore：构建过程或辅助工具的变动
+
+例如：
+
+```bash
+git commit -m "feat: generate converted mov instructions"
+```
+
