+++ include/allinstr.h
@@ -1,0 +1,7 @@
+#ifndef _ALLINSTR_H_
+#define _ALLINSTR_H_ 
+
+void trans_push(char *mode, char *argus[]);
+void trans_push_r(char *reg);
+
+#endif
