+++ tests/dummy.c
@@ -1,0 +1,4 @@
+int main()
+{
+	return 0;
+}
